// javascript/src/openapi/models/Account.ts
/**
 * Wallet OpenAPI Documentation
 * This is a custody wallet service openapi server.
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class Account {
    /**
    * The current balance of the wallet
    */
    'balance': string;
    /**
    * The currency associated with the wallet
    */
    'currency': string;
    /**
    * Account id
    */
    'id': string;
    /**
    * The amount of the balance that is in transit
    */
    'inTransitBalance': string;
    /**
    * The amount of the balance that is locked
    */
    'lockedBalance': string;
    /**
    * The wallet's ID
    */
    'walletId': string;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "balance",
            "baseName": "balance",
            "type": "string",
            "format": ""
        },
        {
            "name": "currency",
            "baseName": "currency",
            "type": "string",
            "format": ""
        },
        {
            "name": "id",
            "baseName": "id",
            "type": "string",
            "format": ""
        },
        {
            "name": "inTransitBalance",
            "baseName": "in_transit_balance",
            "type": "string",
            "format": ""
        },
        {
            "name": "lockedBalance",
            "baseName": "locked_balance",
            "type": "string",
            "format": ""
        },
        {
            "name": "walletId",
            "baseName": "wallet_id",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return Account.attributeTypeMap;
    }
    
    public constructor() {
    }
}