// javascript/src/openapi/models/Address.ts
/**
 * Wallet OpenAPI Documentation
 * This is a custody wallet service openapi server.
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class Address {
    /**
    * the blockchain address.
    */
    'address': string;
    /**
    * the blockchain network
    */
    'network': string;
    /**
    * the type of address.
    */
    'type': string;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "address",
            "baseName": "address",
            "type": "string",
            "format": ""
        },
        {
            "name": "network",
            "baseName": "network",
            "type": "string",
            "format": ""
        },
        {
            "name": "type",
            "baseName": "type",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return Address.attributeTypeMap;
    }
    
    public constructor() {
    }
}