// javascript/src/openapi/models/ChainNetwork.ts
/**
 * Wallet OpenAPI Documentation
 * This is a custody wallet service openapi server.
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class ChainNetwork {
    'explorerAddressUrl'?: string;
    'explorerBlockUrl'?: string;
    'explorerTxUrl'?: string;
    'iconUrl'?: string;
    'id'?: string;
    'name'?: string;
    'network'?: string;
    'networkId'?: string;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "explorerAddressUrl",
            "baseName": "explorer_address_url",
            "type": "string",
            "format": ""
        },
        {
            "name": "explorerBlockUrl",
            "baseName": "explorer_block_url",
            "type": "string",
            "format": ""
        },
        {
            "name": "explorerTxUrl",
            "baseName": "explorer_tx_url",
            "type": "string",
            "format": ""
        },
        {
            "name": "iconUrl",
            "baseName": "icon_url",
            "type": "string",
            "format": ""
        },
        {
            "name": "id",
            "baseName": "id",
            "type": "string",
            "format": ""
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string",
            "format": ""
        },
        {
            "name": "network",
            "baseName": "network",
            "type": "string",
            "format": ""
        },
        {
            "name": "networkId",
            "baseName": "network_id",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return ChainNetwork.attributeTypeMap;
    }
    
    public constructor() {
    }
}