// javascript/src/openapi/models/CreateEndpoint.ts
/**
 * Wallet OpenAPI Documentation
 * This is a custody wallet service openapi server.
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class CreateEndpoint {
    /**
    * The description of the webhook endpoint
    */
    'description'?: string;
    /**
    * The disabled of the webhook endpoint
    */
    'disabled'?: boolean;
    /**
    * The filter event types of the webhook endpoint
    */
    'filterTypes': Array<string>;
    /**
    * The headers of the webhook endpoint
    */
    'headers'?: { [key: string]: string; };
    /**
    * The metadata of the webhook endpoint
    */
    'metadata'?: { [key: string]: any; };
    /**
    * The uid of the webhook endpoint
    */
    'uid'?: string;
    /**
    * The url of the webhook endpoint
    */
    'url': string;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "description",
            "baseName": "description",
            "type": "string",
            "format": ""
        },
        {
            "name": "disabled",
            "baseName": "disabled",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "filterTypes",
            "baseName": "filter_types",
            "type": "Array<string>",
            "format": ""
        },
        {
            "name": "headers",
            "baseName": "headers",
            "type": "{ [key: string]: string; }",
            "format": ""
        },
        {
            "name": "metadata",
            "baseName": "metadata",
            "type": "{ [key: string]: any; }",
            "format": ""
        },
        {
            "name": "uid",
            "baseName": "uid",
            "type": "string",
            "format": ""
        },
        {
            "name": "url",
            "baseName": "url",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return CreateEndpoint.attributeTypeMap;
    }
    
    public constructor() {
    }
}