// javascript/src/openapi/models/CreateExchange.ts
/**
 * Wallet OpenAPI Documentation
 * This is a custody wallet service openapi server.
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class CreateExchange {
    /**
    * The currency code of fiat currency, e.g.: USD
    */
    'fiatCurrency': string;
    /**
    * The amount of the source currency to be exchange, e.g.: 100000
    */
    'fromAmount': string;
    /**
    * The currency code of source currency, e.g.: TON
    */
    'fromCurrency': string;
    /**
    * The currency code of target currency, e.g.: USDT
    */
    'toCurrency': string;
    /**
    * The wallet id
    */
    'walletId': string;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "fiatCurrency",
            "baseName": "fiat_currency",
            "type": "string",
            "format": ""
        },
        {
            "name": "fromAmount",
            "baseName": "from_amount",
            "type": "string",
            "format": ""
        },
        {
            "name": "fromCurrency",
            "baseName": "from_currency",
            "type": "string",
            "format": ""
        },
        {
            "name": "toCurrency",
            "baseName": "to_currency",
            "type": "string",
            "format": ""
        },
        {
            "name": "walletId",
            "baseName": "wallet_id",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return CreateExchange.attributeTypeMap;
    }
    
    public constructor() {
    }
}