// javascript/src/openapi/models/CreateWithdrawRequest.ts
/**
 * Wallet OpenAPI Documentation
 * This is a custody wallet service openapi server.
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class CreateWithdrawRequest {
    /**
    * The amount of currency to be withdrawn
    */
    'amount': string;
    /**
    * The code of currency to be withdrawn.
    */
    'currency': string;
    /**
    * The fee of the withdrawal
    */
    'fee'?: string;
    /**
    * The currency of fee
    */
    'feeCurrency'?: string;
    /**
    * The network of the currency to be withdrawn.
    */
    'network': string;
    /**
    * The recipient address where the withdrawn funds will be sent. This is usually a wallet address specific to the chosen network.
    */
    'toAddress': string;
    /**
    * The ID of the wallet from which the withdrawal will be made.
    */
    'walletId': string;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "amount",
            "baseName": "amount",
            "type": "string",
            "format": ""
        },
        {
            "name": "currency",
            "baseName": "currency",
            "type": "string",
            "format": ""
        },
        {
            "name": "fee",
            "baseName": "fee",
            "type": "string",
            "format": ""
        },
        {
            "name": "feeCurrency",
            "baseName": "fee_currency",
            "type": "string",
            "format": ""
        },
        {
            "name": "network",
            "baseName": "network",
            "type": "string",
            "format": ""
        },
        {
            "name": "toAddress",
            "baseName": "to_address",
            "type": "string",
            "format": ""
        },
        {
            "name": "walletId",
            "baseName": "wallet_id",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return CreateWithdrawRequest.attributeTypeMap;
    }
    
    public constructor() {
    }
}